Create arpd.sh
This bash script that extracts all IP addresses assigned to the system's interfaces using ip addr, retrieves ARP table entries with arp -a, and cross-references the MAC addresses with the IEEE OUI database located at /usr/share/arp-scan/ieee-oui.txt. The script will display the IP addresses assigned to the local system and a table of discovered devices, including their probable vendor.

This script requires ieee-oui.txt to be available (line 12). The document in use is created with the installation of arp-scan. The line may be changed as need to use a local copy of the file.